refactor: add folder structure
