+++ main.ts
@@ -7,15 +7,15 @@ import {
 	PluginSettingTab,
 	addIcon,
 } from "obsidian";
-import Publisher from "./src/Publisher";
-import DigitalGardenSettings from "./src/DigitalGardenSettings";
-import SettingView from "./src/SettingView";
-import { PublishStatusBar } from "./src/PublishStatusBar";
-import { seedling } from "./src/constants";
-import { PublishModal } from "./src/PublishModal";
-import PublishStatusManager from "./src/PublishStatusManager";
-import ObsidianFrontMatterEngine from "./src/ObsidianFrontMatterEngine";
-import DigitalGardenSiteManager from "./src/DigitalGardenSiteManager";
+import Publisher from "./src/publisher/Publisher";
+import DigitalGardenSettings from "./src/models/settings";
+import SettingView from "./src/ui/SettingsView/SettingView";
+import { PublishStatusBar } from "./src/ui/PublishStatusBar";
+import { seedling } from "./src/ui/suggest/constants";
+import { PublishModal } from "./src/ui/PublishModal";
+import PublishStatusManager from "./src/publisher/PublishStatusManager";
+import ObsidianFrontMatterEngine from "./src/publisher/ObsidianFrontMatterEngine";
+import DigitalGardenSiteManager from "./src/publisher/DigitalGardenSiteManager";
 
 const DEFAULT_SETTINGS: DigitalGardenSettings = {
 	githubRepo: "",
