+++ src/models/pluginInfo.ts
