+++ src/publisher/Validator.ts
