+++ src/ui/PublishStatusBar.ts
